redefine how schema is defined
