+++ apps/kdx/src/app/[locale]/team/notifications/preview-email/[id]/route.ts
@@ -3,7 +3,7 @@ import { notFound, redirect } from "next/navigation";
 import { auth } from "@kdx/auth";
 import { eq, sql } from "@kdx/db";
 import { db } from "@kdx/db/client";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 const allTeamIdsForUserQuery = db
   .select({ id: schema.usersToTeams.teamId })
