+++ packages/api/src/internal/notificationCenter.ts
@@ -1,7 +1,7 @@
 import { render } from "@react-email/render";
 
 import { db } from "@kdx/db/client";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { resend } from "../utils/email";
