+++ packages/api/src/routers/app/calendar/getAll.handler.ts
@@ -3,7 +3,7 @@ import { rrulestr } from "rrule";
 import type { TGetAllInputSchema } from "@kdx/validators/trpc/app/calendar";
 import dayjs from "@kdx/dayjs";
 import { and, eq, gte, lte, or } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TProtectedProcedureContext } from "../../../procedures";
