+++ packages/api/src/routers/app/calendar/nuke.handler.ts
@@ -1,7 +1,7 @@
 import { TRPCError } from "@trpc/server";
 
 import { eq } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 import { authorizedEmails } from "@kdx/shared";
 
 import type { TProtectedProcedureContext } from "../../../procedures";
