+++ packages/api/src/routers/app/installApp.handler.ts
@@ -4,7 +4,7 @@ import type { TInstallAppInputSchema } from "@kdx/validators/trpc/app";
 import { and, eq } from "@kdx/db";
 import { appRoles_defaultTree } from "@kdx/db/constants";
 import { nanoid } from "@kdx/db/nanoid";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 import { appIdToAdminRole_defaultIdMap } from "@kdx/shared";
 
 import type { TIsTeamOwnerProcedureContext } from "../../procedures";
