+++ packages/api/src/routers/app/kodixCare/doCheckoutForShift.handler.ts
@@ -3,7 +3,7 @@ import { TRPCError } from "@trpc/server";
 import type { TDoCheckoutForShiftInputSchema } from "@kdx/validators/trpc/app/kodixCare";
 import dayjs from "@kdx/dayjs";
 import { eq } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TProtectedProcedureContext } from "../../../procedures";
 import { getCurrentCareShiftHandler } from "./getCurrentCareShift.handler";
