+++ packages/api/src/routers/app/kodixCare/getCareTasks.handler.ts
@@ -1,6 +1,6 @@
 import type { TGetCareTasksInputSchema } from "@kdx/validators/trpc/app/kodixCare";
 import dayjs from "@kdx/dayjs";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 import { kodixCareAppId } from "@kdx/shared";
 
 import type { TProtectedProcedureContext } from "../../../procedures";
