+++ packages/api/src/routers/app/kodixCare/saveCareTask.handler.ts
@@ -1,6 +1,6 @@
 import type { TSaveCareTaskInputSchema } from "@kdx/validators/trpc/app/kodixCare";
 import { eq } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TProtectedProcedureContext } from "../../../procedures";
