+++ packages/api/src/routers/team/appRole/getPermissions.handler.ts
@@ -1,6 +1,6 @@
 import type { TGetPermissionsInputSchema } from "@kdx/validators/trpc/team/appRole";
 import { eq } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TIsTeamOwnerProcedureContext } from "../../../procedures";
