+++ packages/api/src/routers/team/appRole/updatePermissionAssociation.handler.ts
@@ -1,6 +1,6 @@
 import type { TUpdatePermissionAssociationInputSchema } from "@kdx/validators/trpc/team/appRole";
 import { eq, inArray } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TIsTeamOwnerProcedureContext } from "../../../procedures";
