+++ packages/api/src/routers/team/getAllUsers.handler.ts
@@ -1,5 +1,5 @@
 import { eq } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TProtectedProcedureContext } from "../../procedures";
