+++ packages/api/src/routers/team/invitation/delete.handler.ts
@@ -2,7 +2,7 @@ import { TRPCError } from "@trpc/server";
 
 import type { TDeleteInputSchema } from "@kdx/validators/trpc/team/invitation";
 import { and, eq } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TProtectedProcedureContext } from "../../../procedures";
