+++ packages/api/src/routers/team/invitation/invite.handler.ts
@@ -3,7 +3,7 @@ import { TRPCError } from "@trpc/server";
 
 import type { TInviteInputSchema } from "@kdx/validators/trpc/team/invitation";
 import { nanoid } from "@kdx/db/nanoid";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 import TeamInvite from "@kdx/react-email/team-invite";
 import {
   getBaseUrl,
