+++ packages/api/src/routers/team/removeUser.handler.ts
@@ -3,7 +3,7 @@ import { TRPCError } from "@trpc/server";
 import type { TRemoveUserSchema } from "@kdx/validators/trpc/team";
 import { and, eq, not } from "@kdx/db";
 import { nanoid } from "@kdx/db/nanoid";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TProtectedProcedureContext } from "../../procedures";
