+++ packages/api/src/routers/team/update.handler.ts
@@ -1,6 +1,6 @@
 import type { TUpdateInputSchema } from "@kdx/validators/trpc/team";
 import { eq } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TProtectedProcedureContext } from "../../procedures";
