+++ packages/api/src/routers/user/sendResetPasswordEmail.ts
@@ -3,7 +3,7 @@ import { TRPCError } from "@trpc/server";
 import type { TSendResetPasswordEmailInputSchema } from "@kdx/validators/trpc/user";
 import { eq } from "@kdx/db";
 import { nanoid } from "@kdx/db/nanoid";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 import ResetPassword from "@kdx/react-email/reset-password";
 import { kodixNotificationFromEmail } from "@kdx/shared";
