+++ packages/api/src/routers/user/signupWithPassword.handler.ts
@@ -8,7 +8,7 @@ import { createUser } from "@kdx/auth/db";
 import { eq } from "@kdx/db";
 import { db } from "@kdx/db/client";
 import { nanoid } from "@kdx/db/nanoid";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TPublicProcedureContext } from "../../procedures";
 import { argon2Config } from "./utils";
