+++ packages/api/src/routers/user/switchActiveTeam.handler.ts
@@ -1,6 +1,6 @@
 import type { TSwitchActiveTeamInputSchema } from "@kdx/validators/trpc/user";
 import { eq } from "@kdx/db";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import type { TProtectedProcedureContext } from "../../procedures";
