+++ packages/api/src/upstash.ts
@@ -1,6 +1,6 @@
 import { Redis } from "@upstash/redis";
 
-import type { schema } from "@kdx/db/schema";
+import type { apps, teams, users } from "@kdx/db/schema";
 import type { AppPermissionId } from "@kdx/shared";
 
 const redis = new Redis({
@@ -11,19 +11,19 @@ const redis = new Redis({
 interface KeysMapping {
   apps: {
     tags: {
-      teamId: typeof schema.teams.$inferSelect.id | undefined; //? If teamId isn't provided, it will refer to all existant apps
+      teamId: typeof teams.$inferSelect.id | undefined; //? If teamId isn't provided, it will refer to all existant apps
     };
     value: {
-      id: typeof schema.apps.$inferSelect.id;
+      id: typeof apps.$inferSelect.id;
       installed: boolean;
     }[];
   };
 
   permissions: {
     tags: {
-      userId: typeof schema.users.$inferSelect.id;
+      userId: typeof users.$inferSelect.id;
       permissionId: AppPermissionId;
-      teamId: typeof schema.teams.$inferSelect.id;
+      teamId: typeof teams.$inferSelect.id;
     };
     value:
       | {
