+++ packages/auth/src/providers/utils/createOrGetExistingUserForUnlinkedProviderAccount.ts
@@ -2,7 +2,7 @@ import { cookies } from "next/headers";
 
 import { db } from "@kdx/db/client";
 import { nanoid } from "@kdx/db/nanoid";
-import { schema } from "@kdx/db/schema";
+import * as schema from "@kdx/db/schema";
 
 import { createUser } from "../../db";
