+++ packages/db/src/schema/index.ts
@@ -1,33 +1,6 @@
-import type {
-  MySqlTableWithColumns,
-  TableConfig,
-} from "drizzle-orm/mysql-core";
-
-import { calendarAppId, kodixCareAppId, todoAppId } from "@kdx/shared";
-
-import * as apps from "./apps";
-import * as calendar from "./apps/calendar";
-import * as kodixCare from "./apps/kodixCare";
-import * as todos from "./apps/todos";
-import * as teams from "./teams";
-import * as users from "./users";
-
-export const schema = {
-  ...apps,
-  ...calendar,
-  ...kodixCare,
-  ...todos,
-  ...users,
-  ...teams,
-};
-
-const withoutRelations = <T>(obj: Record<string, T>) =>
-  Object.fromEntries(
-    Object.entries(obj).filter(([key]) => !key.includes("Relation")),
-  ) as Record<string, MySqlTableWithColumns<TableConfig>>;
-
-export const appIdToSchemas = {
-  [kodixCareAppId]: withoutRelations(kodixCare),
-  [todoAppId]: withoutRelations(todos),
-  [calendarAppId]: withoutRelations(calendar),
-};
+export * from "./apps";
+export * from "./apps/calendar";
+export * from "./apps/kodixCare";
+export * from "./apps/todos";
+export * from "./teams";
+export * from "./users";
