+++ packages/validators/src/trpc/app/todo/index.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import type { schema } from "@kdx/db/schema";
+import type { todos } from "@kdx/db/schema";
 
 import { ZNanoId } from "../../..";
 
@@ -9,7 +9,7 @@ export const ZCreateInputSchema = z.object({
   description: z.string().optional(),
   dueDate: z.date().optional(),
   priority: z.number().optional(),
-  status: z.custom<typeof schema.todos.$inferInsert.status>().optional(),
+  status: z.custom<typeof todos.$inferInsert.status>().optional(),
   assignedToUserId: ZNanoId.optional().nullish(),
 });
 export type TCreateInputSchema = z.infer<typeof ZCreateInputSchema>;
@@ -20,7 +20,7 @@ export const ZUpdateInputSchema = z.object({
   description: z.string().optional(),
   dueDate: z.date().optional().nullish(),
   priority: z.number().optional(),
-  status: z.custom<typeof schema.todos.$inferInsert.status>().optional(),
+  status: z.custom<typeof todos.$inferInsert.status>().optional(),
   assignedToUserId: ZNanoId.optional().nullish(),
 }); //TODO: investigate
 export type TUpdateInputSchema = z.infer<typeof ZUpdateInputSchema>;
